70% Yeah

+++ pages/dashboard/tasks_list.php
@@ -1,5 +1,18 @@
 <?php
+session_start();
+
+// Check the session
+if (!isset($_SESSION['login'])) {
+    header('Location: login.php');
+    exit;
+}
+
 include './../../includes/function.php';
+
+// Get user's session data
+$email = $_SESSION['email'];
+$user = query("SELECT DISTINCT * FROM tb_users WHERE email = '$email'");
+$kdUser = $user[0]['kd_user'];
 
 // Pagination Set up
 $page = isset($_GET['page']) ? (int)$_GET['page'] : 1;
@@ -8,10 +21,19 @@ $perPage = 5;
 $start = ($page - 1) * $perPage;
 
 // Ambil data sesuai halaman
-$assignments = query("SELECT * FROM tb_tugas ORDER BY created_at DESC LIMIT $start, $perPage");
+if ($user[0]['role'] == 'Penjoki') {
+    $assignments = query("SELECT * FROM tb_tugas WHERE kd_penjoki = '$kdUser' ORDER BY created_at DESC LIMIT $start, $perPage");
+    $total = count(query("SELECT * FROM tb_tugas WHERE kd_penjoki = '$kdUser'"));
+} else if ($user[0]['role'] == 'Pengguna') {
+    $assignments = query("SELECT * FROM tb_tugas WHERE kd_user = '$kdUser' ORDER BY created_at DESC LIMIT $start, $perPage");
+    $total = count(query("SELECT * FROM tb_tugas WHERE kd_user = '$kdUser'"));
+} else {
+    $assignments = query("SELECT * FROM tb_tugas ORDER BY created_at DESC LIMIT $start, $perPage");
+    $total = count(query("SELECT * FROM tb_tugas"));
+}
+
 
 // Hitung total data
-$total = count(query("SELECT * FROM tb_tugas"));
 $totalPages = ceil($total / $perPage);
 ?>
 <!DOCTYPE html>
@@ -29,6 +51,7 @@ $totalPages = ceil($total / $perPage);
     <link href="https://unpkg.com/aos@2.3.1/dist/aos.css" rel="stylesheet">
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/animate.css/4.1.1/animate.min.css" />
     <title>Daftar Tugas | Jokiin</title>
+
 </head>
 
 <body>
@@ -42,10 +65,12 @@ $totalPages = ceil($total / $perPage);
             <i class="ri-dashboard-line" onclick="location.href='dashboard.php'"></i>
             <a href="./dashboard.php">Dashboard</a>
         </div>
-        <div class="box">
-            <i class="ri-group-line" onclick="location.href='user_management.php'"></i>
-            <a href="./user_management.php">Manajemen Pengguna</a>
-        </div>
+        <?php if ($user[0]['role'] == 'Admin') : ?>
+            <div class="box">
+                <i class="ri-group-line" onclick="location.href='user_management.php'"></i>
+                <a href="./user_management.php">Manajemen Pengguna</a>
+            </div>
+        <?php endif; ?>
         <div class="box">
             <i class="ri-message-2-line" onclick="location.href='mail_management.php'"></i>
             <a href="./mail_management.php">Manajemen Pesan</a>
@@ -63,12 +88,12 @@ $totalPages = ceil($total / $perPage);
             <a href="./rating.php">Penilaian</a>
         </div>
         <div class="box">
-            <i class="ri-home-5-line" onclick="location.href='./../home.html'"></i>
-            <a href="./../home.html">Beranda</a>
+            <i class="ri-home-5-line" onclick="location.href='./../home.php'"></i>
+            <a href="./../home.php">Beranda</a>
         </div>
         <div class="box">
-            <i class="ri-logout-circle-line" onclick="location.href='./../../index.html'"></i>
-            <a href="./../../index.html">Keluar</a>
+            <i class="ri-logout-circle-line" onclick="location.href='./../logout.php'"></i>
+            <a href="./../logout.php">Keluar</a>
         </div>
     </section>
     <!-- Sidebar End -->
@@ -87,7 +112,7 @@ $totalPages = ceil($total / $perPage);
                             <input type="text" name="cari" id="cari" placeholder="Cari sesuatu">
                         </form>
                         <img src="./../../img/user-1.jpg" alt="user-1">
-                        <h2>John Nash</h2>
+                        <h2><?= $user[0]['nama_lengkap']; ?></h2>
                     </div>
                 </div>
             </div>
@@ -128,29 +153,37 @@ $totalPages = ceil($total / $perPage);
                                     <th>Status</th>
                                     <th>Aksi</th>
                                 </tr>
-                                <?php $no = $start + 1; ?>
-                                <?php foreach ($assignments as $assignment) : ?>
+                                <?php if (!empty($assignments)): ?>
+                                    <?php $no = $start + 1; ?>
+                                    <?php foreach ($assignments as $assignment) : ?>
+                                        <tr>
+                                            <td><?= $no++;; ?></td>
+                                            <td><?= $assignment['kd_tugas']; ?></td>
+                                            <td><?= $assignment['kd_user']; ?></td>
+                                            <td><?= $assignment['kd_penjoki']; ?></td>
+                                            <td class="desc"><strong><?= $assignment['judul']; ?></strong></td>
+                                            <td class="desc"><?= $assignment['deskripsi']; ?></td>
+                                            <td><?= $assignment['assignment_type']; ?></td>
+                                            <td><?= $assignment['created_at']; ?></td>
+                                            <td>
+                                                <div class="status <?= $assignment['is_finished'] == 'false' ? '' : ' finished'; ?>">
+                                                    <?= $assignment['is_finished'] == 'false' ? 'Belum selesai' : 'Selesai'; ?>
+                                                </div>
+                                            </td>
+                                            <td class="button-action">
+                                                <button class="warning" onclick="location.href='./edit_consultant.php?kd_tugas=<?= $assignment['kd_tugas']; ?>'"><i class="ri-pencil-line"></i></button>
+                                                <?php if ($user[0]['role'] == 'Pengguna'): ?>
+                                                    <button class="danger" onclick="location.href='./delete_task.php?kd_tugas=<?= $assignment['kd_tugas']; ?>'"><i class="ri-delete-bin-line"></i></button>
+                                                <?php endif; ?>
+                                                <button onclick="location.href='./../detail.php?kd_tugas=<?= $assignment['kd_tugas']; ?>'"><i class="ri-gallery-view-2"></i></button>
+                                            </td>
+                                        </tr>
+                                    <?php endforeach; ?>
+                                <?php else : ?>
                                     <tr>
-                                        <td><?= $no++;; ?></td>
-                                        <td><?= $assignment['kd_tugas']; ?></td>
-                                        <td><?= $assignment['kd_user']; ?></td>
-                                        <td><?= $assignment['kd_penjoki']; ?></td>
-                                        <td class="desc"><strong><?= $assignment['judul']; ?></strong></td>
-                                        <td class="desc"><?= $assignment['deskripsi']; ?></td>
-                                        <td><?= $assignment['assignment_type']; ?></td>
-                                        <td><?= $assignment['created_at']; ?></td>
-                                        <td>
-                                            <div class="status <?= $assignment['is_finished'] == 'false' ? '' : ' finished'; ?>">
-                                                <?= $assignment['is_finished'] == 'false' ? 'Belum selesai' : 'Selesai'; ?>
-                                            </div>
-                                        </td>
-                                        <td class="button-action">
-                                            <button class="warning" onclick="location.href='./edit_consultant.php?kd_tugas=<?= $assignment['kd_tugas']; ?>'"><i class="ri-pencil-line"></i></button>
-                                            <button class="danger" onclick="location.href='./delete_task.php?kd_tugas=<?= $assignment['kd_tugas']; ?>'"><i class="ri-delete-bin-line"></i></button>
-                                            <button onclick="location.href='./../detail.html'"><i class="ri-gallery-view-2"></i></button>
-                                        </td>
+                                        <td colspan="10">Tidak memiliki data</td>
                                     </tr>
-                                <?php endforeach; ?>
+                                <?php endif; ?>
                             </table>
                         </div>
                     </div>
@@ -190,6 +223,8 @@ $totalPages = ceil($total / $perPage);
     <!-- Main-app -->
 
     <script src="./../../dist/js/script.js"></script>
+    <script src="https://cdn.jsdelivr.net/npm/jquery@3.5.1/dist/jquery.slim.min.js" integrity="sha384-DfXdz2htPH0lsSSs5nCTpuj/zy4C+OGpamoFVy38MVBnE+IbbVYUew+OrCXaRkfj" crossorigin="anonymous"></script>
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@4.6.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-Fy6S3B9q64WdZWQUiU+q4/2Lc9npb8tCaSX9FK7E8HnRr0Jz8D6OP9dO5Vg3Q9ct" crossorigin="anonymous"></script>
 </body>
 
 </html>
